Created Base object for Simulation object to separate cell array setup methods and other variables
